Change some types to checked_int

+++ components/core/wf/code_generation/ast_conversion.cc
@@ -422,7 +422,8 @@ ast::ast_element ast_form_visitor::operator()(const ir::value&, const ir::load& 
         if constexpr (std::is_same_v<T, symbolic_constant>) {
           return ast::ast_element{std::in_place_type_t<ast::special_constant>{}, inner.name()};
         } else if constexpr (std::is_same_v<T, integer_constant>) {
-          return ast::ast_element{std::in_place_type_t<ast::integer_literal>{}, inner.get_value()};
+          return ast::ast_element{std::in_place_type_t<ast::integer_literal>{},
+                                  static_cast<std::int64_t>(inner.get_value())};
         } else if constexpr (std::is_same_v<T, float_constant>) {
           return ast::ast_element{std::in_place_type_t<ast::float_literal>{},
                                   static_cast<float_constant>(inner).get_value()};
